[Streaming, Preview] Remove `watershed`, use `ws` in NodeWebSocket (#1425)
* remove watershed socket impl, work on tests

* clean-up bf-streaming package.json, add eslint config

* cleanup BFAdapter streaming tests

* clean up casing on relative imports

* apply PR feedback to NodeWebSocket class
