+++ libraries/botframework-streaming/src/webSocket/factories/index.ts
@@ -8,4 +8,3 @@
 
 export * from './nodeWebSocketFactory';
 export * from './nodeWebSocketFactoryBase';
-export * from './wsNodeWebSocketFactory';
