+++ libraries/botframework-streaming/src/webSocket/factories/nodeWebSocketFactoryBase.ts
@@ -11,5 +11,5 @@ import { Socket } from 'net';
 import { ISocket } from '../../interfaces';
 
 export abstract class NodeWebSocketFactoryBase {
-    public abstract createWebSocket(req: IncomingMessage, socket: Socket, head: Buffer): ISocket;
+    public abstract createWebSocket(req: IncomingMessage, socket: Socket, head: Buffer): Promise<ISocket>;
 }
